Added two new test cases for numeric functions. Also includes a few bug fixes and clean up items.

+++ vxquery-core/src/main/java/org/apache/vxquery/datamodel/accessors/atomic/XSDecimalPointable.java
@@ -197,7 +197,7 @@ public class XSDecimalPointable extends AbstractPointable implements IHashable, 
 
     @Override
     public double doubleValue() {
-        return getDecimalValue() / Math.pow(10, getDecimalPlace());
+        return ((double) getDecimalValue()) / Math.pow(10, getDecimalPlace());
     }
 
 }
